requirements.txt
Django==3.1.7

django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djoser==2.1.0
social-auth-app-django==4.0.0

django-ckeditor==6.1.0

pytz==2021.3

Pillow==9.5.0
mutagen==1.45.1 

pymysql==1.1.0

argon2-cffi==21.1.0 
django-environ==0.4.5

stripe==2.58.0
braintree==4.13.1

pyJWT==2.1.0

boto3==1.18.63
django-storages==1.12.2 

markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.3.0

requests==2.26.0 
